+++ include/roq/mask.hpp
@@ -7,7 +7,6 @@
 #include <utility>
 
 namespace roq {
-namespace utils {
 
 // implements a bitmap mask using an underlying enumeration
 // some limitations:
@@ -125,5 +124,4 @@ class Mask final {
   value_type value_ = {};
 };
 
-}  // namespace utils
 }  // namespace roq